int to size_t

+++ main.c
@@ -12,34 +12,34 @@ int main(int argc, char *argv[]){
     zstring removedCharString = newZString("");
     zstring reversedString = newZString("");
     
-    int occurances = 0;
+    size_t  occurances = 0;
     char toFind = 'w';
     char toRemove[] = "world";
     Bool areEqual = False;
 
     str.length = lenOfStr(str.data);
-    printf("String: '%s' of Len: %d\n", str.data, str.length);
+    printf("String: '%s' of Len: %zu\n", str.data, str.length);
 
     trimmedString = trimStr(str);
-    printf("Trimmed: '%s' of len: %d\n", trimmedString.data, trimmedString.length);
+    printf("Trimmed: '%s' of len: %zu\n", trimmedString.data, trimmedString.length);
     
     occurances = findOccuranceOf(str, toFind);
-    printf("In the string: '%s' the '%c' occures %d times\n", str.data, toFind, occurances);
+    printf("In the string: '%s' the '%c' occures %zu times\n", str.data, toFind, occurances);
 
     removedWordString  = removeWord(str, toRemove);
-    printf("Removed '%s' at index %d to String: '%s' of len: %d\n", toRemove, findStartOfWord(str, toRemove),removedWordString.data, removedWordString.length);
+    printf("Removed '%s' at index %zu to String: '%s' of len: %zu\n", toRemove, findStartOfWord(str, toRemove),removedWordString.data, removedWordString.length);
     
     copiedString = copyStr(str);
-    printf("Copied: '%s' of len: %d\n", copiedString.data, copiedString.length);
+    printf("Copied: '%s' of len: %zu\n", copiedString.data, copiedString.length);
     
     areEqual = compareString(str, toLowercaseStr(str));
     printf("String 1 : '%s' and String 2: '%s' are equal? %s\n", str.data, toLowercaseStr(str).data, areEqual ? "True" : "False");
     
     removedCharString  = removeChar(str, toFind);
-    printf("Removed '%c' to String: '%s' of len: %d\n", toFind, removedCharString.data, removedCharString.length);
+    printf("Removed '%c' to String: '%s' of len: %zu\n", toFind, removedCharString.data, removedCharString.length);
 
     reversedString = reverseStr(str);
-    printf("Reversed: '%s' of len %d\n", reversedString.data, reversedString.length);
+    printf("Reversed: '%s' of len %zu\n", reversedString.data, reversedString.length);
 
     return 0;
 }
